refactor: convert images to webp, reduce size of images
