+++ src/pages/PageNotFound/constants.ts
@@ -2,7 +2,7 @@ import { Paths } from '@/routes/routeConstants';
 
 export enum InfoCardData {
   heading = 'Oops!',
-  imgPath = '/images/notFound/notFound.png',
+  imgPath = '/images/notFound/notFound.webp',
   subHeading = 'The page you requested was not found!',
 }
